Move stuff related to extensions to @directus/extensions
