+++ packages/extensions/package.json
@@ -1,0 +1,50 @@
+{
+	"name": "@directus/extensions",
+	"version": "0.0.1",
+	"description": "Utilities and Types for extensions",
+	"homepage": "https://directus.io",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/directus/directus.git",
+		"directory": "packages/extensions"
+	},
+	"funding": "https://github.com/directus/directus?sponsor=1",
+	"license": "MIT",
+	"author": "Rijk van Zanten <rijkvanzanten@me.com>",
+	"type": "module",
+	"exports": {
+		".": "./dist/index.js",
+		"./package.json": "./package.json"
+	},
+	"main": "dist/index.js",
+	"files": [
+		"dist"
+	],
+	"scripts": {
+		"build": "tsup src/index.ts --format=esm --dts",
+		"dev": "tsup src/index.ts --format=esm --dts --watch",
+		"test": "vitest --watch=false"
+	},
+	"devDependencies": {
+		"@directus/constants": "workspace:*",
+		"@directus/tsconfig": "workspace:*",
+		"@directus/types": "workspace:*",
+		"@types/express": "4.17.17",
+		"@types/fs-extra": "11.0.1",
+		"@types/tmp": "0.2.3",
+		"@vitest/coverage-c8": "0.31.1",
+		"knex": "2.5.1",
+		"pino": "8.14.1",
+		"tsup": "7.2.0",
+		"typescript": "5.2.2",
+		"vitest": "0.31.1",
+		"vue": "3.3.4",
+		"vue-router": "4.2.0"
+	},
+	"dependencies": {
+		"@directus/utils": "workspace:*",
+		"fs-extra": "11.1.1",
+		"tmp": "0.2.1",
+		"zod": "3.21.4"
+	}
+}
