+++ packages/extensions/src/constants/index.ts
@@ -1,0 +1,5 @@
+export * from './extension-types.js';
+export * from './languages.js';
+export * from './name-regex.js';
+export * from './pkg-key.js';
+export * from './shared-deps.js';
