+++ packages/extensions/src/constants/languages.ts
@@ -1,0 +1,1 @@
+export const EXTENSION_LANGUAGES = ['javascript', 'typescript'] as const;
