+++ packages/extensions/src/constants/name-regex.ts
@@ -1,0 +1,1 @@
+export const EXTENSION_NAME_REGEX = /^(?:(?:@[^/]+\/)?directus-extension-|@directus\/extension-)(.+)$/;
