+++ packages/extensions/src/constants/pkg-key.ts
@@ -1,0 +1,1 @@
+export const EXTENSION_PKG_KEY = 'directus:extension';
