+++ packages/extensions/src/constants/shared-deps.ts
@@ -1,0 +1,3 @@
+export const APP_SHARED_DEPS = ['@directus/extensions-sdk', 'vue', 'vue-router', 'vue-i18n', 'pinia'];
+
+export const API_SHARED_DEPS = ['directus'];
