+++ packages/extensions/src/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from './manifest.js';
+export * from './options.js';
