+++ packages/extensions/src/types/displays.ts
@@ -1,7 +1,6 @@
+import type { DeepPartial, Field, LocalType, Type } from '@directus/types';
 import type { Component, ComponentOptions } from 'vue';
-import type { ExtensionOptionsContext } from './extensions.js';
-import type { Field, LocalType, Type } from './fields.js';
-import type { DeepPartial } from './misc.js';
+import type { ExtensionOptionsContext } from './options.js';
 
 export type DisplayFieldsFunction = (
 	options: any,
