+++ packages/extensions/src/types/info.ts
@@ -1,0 +1,7 @@
+import type { ApiExtension, AppExtension, BundleExtension, HybridExtension } from './extension-types.js';
+
+export type ExtensionInfo =
+	| Omit<AppExtension, 'entrypoint' | 'path'>
+	| Omit<ApiExtension, 'entrypoint' | 'path'>
+	| Omit<HybridExtension, 'entrypoint' | 'path'>
+	| Omit<BundleExtension, 'entrypoint' | 'path'>;
