+++ packages/extensions/src/types/modules.ts
@@ -1,6 +1,5 @@
+import type { Permission, User } from '@directus/types';
 import type { RouteRecordRaw } from 'vue-router';
-import type { Permission } from './permissions.js';
-import type { User } from './users.js';
 
 export interface ModuleConfig {
 	id: string;
