+++ packages/extensions/src/types/operations.ts
@@ -1,9 +1,6 @@
+import type { Accountability, DeepPartial, Field, FlowRaw } from '@directus/types';
 import type { ComponentOptions } from 'vue';
-import type { Accountability } from './accountability.js';
-import type { ApiExtensionContext } from './extensions.js';
-import type { Field } from './fields.js';
-import type { DeepPartial } from './misc.js';
-import type { FlowRaw } from './flows.js';
+import type { ApiExtensionContext } from './api-extension-context.js';
 
 export type OperationContext = ApiExtensionContext & {
 	data: Record<string, unknown>;
