+++ packages/extensions/src/types/options.ts
@@ -1,0 +1,29 @@
+import type { LOCAL_TYPES } from '@directus/constants';
+import type { Collection, DeepPartial, Field, Relation } from '@directus/types';
+
+export type ExtensionOptionsContext = {
+	collection: string | undefined;
+	editing: string;
+	field: DeepPartial<Field>;
+	relations: {
+		m2o: DeepPartial<Relation> | undefined;
+		m2a?: DeepPartial<Relation> | undefined;
+		o2m: DeepPartial<Relation> | undefined;
+	};
+	collections: {
+		junction: DeepPartial<Collection & { fields: DeepPartial<Field>[] }> | undefined;
+		related: DeepPartial<Collection & { fields: DeepPartial<Field>[] }> | undefined;
+	};
+	fields: {
+		corresponding: DeepPartial<Field> | undefined;
+		junctionCurrent: DeepPartial<Field> | undefined;
+		junctionRelated: DeepPartial<Field> | undefined;
+		sort: DeepPartial<Field> | undefined;
+	};
+
+	items: Record<string, Record<string, any>[]>;
+
+	localType: (typeof LOCAL_TYPES)[number];
+	autoGenerateJunctionRelation: boolean;
+	saving: boolean;
+};
