+++ packages/extensions/src/types/panels.ts
@@ -1,7 +1,5 @@
+import type { DeepPartial, Field, Query } from '@directus/types';
 import type { Component, ComponentOptions } from 'vue';
-import type { DeepPartial } from './misc.js';
-import type { Field } from './fields.js';
-import type { Query } from './query.js';
 
 export type PanelQuery = { collection: string; query: Query; key?: string };
