+++ packages/extensions/src/utils/ensure-extension-dirs.test.ts
