+++ packages/extensions/src/utils/generate-extensions-entrypoint.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
-
-import type { Extension } from '@directus/types';
+import type { Extension } from '../types/index.js';
 import { generateExtensionsEntrypoint } from './generate-extensions-entrypoint.js';
 
 describe('generateExtensionsEntrypoint', () => {
