+++ packages/extensions/src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './ensure-extension-dirs.js';
+export * from './generate-extensions-entrypoint.js';
+export * from './get-extensions.js';
