+++ packages/types/src/index.ts
@@ -1,23 +1,14 @@
 export * from './accountability.js';
 export * from './collection.js';
-export * from './displays.js';
-export * from './endpoints.js';
 export * from './events.js';
-export * from './extensions.js';
 export * from './fields.js';
 export * from './files.js';
 export * from './filter.js';
 export * from './flows.js';
 export * from './geometry.js';
-export * from './hooks.js';
-export * from './interfaces.js';
 export * from './items.js';
-export * from './layouts.js';
 export * from './misc.js';
-export * from './modules.js';
 export * from './notifications.js';
-export * from './operations.js';
-export * from './panels.js';
 export * from './permissions.js';
 export * from './presets.js';
 export * from './query.js';
