+++ packages/utils/node/index.ts
@@ -1,7 +1,4 @@
 export * from './array-helpers.js';
-export * from './ensure-extension-dirs.js';
-export * from './generate-extensions-entrypoint.js';
-export * from './get-extensions.js';
 export * from './is-readable-stream.js';
 export * from './list-folders.js';
 export * from './path-to-relative-url.js';
